+++ libs/exavmlib/lib/String.Chars.Float.ex
@@ -1,8 +1,8 @@
 #
 # This file is part of elixir-lang.
 #
-# Copyright 2012-2024 Elixir Contributors
-# https://github.com/elixir-lang/elixir/tree/v1.17.2/lib/elixir/lib/map_set.ex
+# Copyright 2013-2023 Elixir Contributors
+# https://github.com/elixir-lang/elixir/commits/v1.17.2/lib/elixir/lib/string/chars.ex
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,21 +19,11 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-defimpl Enumerable, for: MapSet do
-  def count(map_set) do
-    {:ok, MapSet.size(map_set)}
-  end
+import Kernel, except: [to_string: 1]
 
-  def member?(map_set, val) do
-    {:ok, MapSet.member?(map_set, val)}
-  end
-
-  def reduce(map_set, acc, fun) do
-    Enumerable.List.reduce(MapSet.to_list(map_set), acc, fun)
-  end
-
-  def slice(map_set) do
-    size = MapSet.size(map_set)
-    {:ok, size, &Enumerable.List.slice(MapSet.to_list(map_set), &1, &2, size)}
+defimpl String.Chars, for: Float do
+  def to_string(term) do
+    # TODO: :short option not yet supported right now, so :decimals+:compact should be replaced
+    :erlang.float_to_binary(term, [{:decimals, 17}, :compact])
   end
 end